In the protocol, the length field changed its size from uint8_t to uint16_t. OTA in progress

+++ unwired/common-node.h
@@ -80,7 +80,7 @@ volatile union
 void pack_sender(const uip_ip6addr_t *dest_addr, 
 				 uint8_t device_id, 
 				 uint8_t data_type, 
-				 uint8_t payload_len, 
+				 uint16_t payload_len, 
 				 uint8_t *payload);
 
 /* Функция управления светодиодами */
